requirements.txt
streamlit
torch
transformers
pillow
openai-whisper  
pyttsx3
audio-recorder-streamlit